Refactored work into separate classes
