+++ src/DotNetCensus.Core/GlobalSuppressions.cs
@@ -7,4 +7,6 @@ using System.Diagnostics.CodeAnalysis;
 
 [assembly: SuppressMessage("Style", "IDE1006:Naming Styles", Justification = "Is GitHub class", Scope = "member", Target = "~P:DotNetCensus.Core.Models.GitHub.TreeResponse.tree")]
 [assembly: SuppressMessage("Style", "IDE1006:Naming Styles", Justification = "Is GitHub class", Scope = "member", Target = "~P:DotNetCensus.Core.Models.GitHub.FileResponse.path")]
-[assembly: SuppressMessage("Style", "IDE0074:Use compound assignment", Justification = "Rule doesn't improve readability", Scope = "member", Target = "~M:DotNetCensus.Core.ProjectScanning.SearchProjectFile(System.IO.FileInfo,System.String,System.IO.FileInfo)~System.Collections.Generic.List{DotNetCensus.Core.Models.Project}")]
+[assembly: SuppressMessage("Style", "IDE1006:Naming Styles", Justification = "Is GitHub class", Scope = "member", Target = "~P:DotNetCensus.Core.Models.GitHub.FileDetails.type")]
+[assembly: SuppressMessage("Style", "IDE1006:Naming Styles", Justification = "Is GitHub class", Scope = "member", Target = "~P:DotNetCensus.Core.Models.GitHub.FileDetails.size")]
+//[assembly: SuppressMessage("Style", "IDE0074:Use compound assignment", Justification = "Rule doesn't improve readability", Scope = "member", Target = "~M:DotNetCensus.Core.DirectoryProjectScanning.SearchProjectFile(System.IO.FileInfo,System.String,System.IO.FileInfo)~System.Collections.Generic.List{DotNetCensus.Core.Models.Project}")]
