+++ src/DotNetCensus.Core/Models/GitHub/RepoDirectory.cs
@@ -2,8 +2,8 @@
 
 public class RepoDirectory
 {
-    public string Name { get; set; }
-    public string Path { get; set; }
+    public string? Name { get; set; }
+    public string? Path { get; set; }
     public List<RepoDirectory> Directories { get; set; } = new();
     public List<string> Files { get; set; } = new();
 }
